Make highlands auth working
We needed to have the classical rails autoloader that we were not doing, this was giving us the error ```The action 'new' could not be found for HighlandsAuth::SessionsController```


Also, we needed to have support to haml that has been fixed the support recently in rails 6. Basicaly those were the two major changes to make everything working.